refactor: moved validationResult back to @taquito/core
