+++ packages/taquito-core/src/taquito-core.ts
@@ -4,4 +4,3 @@
  */
 
 export * from './errors';
-export * from './enums';
