+++ packages/taquito-local-forging/src/codec.ts
@@ -7,6 +7,7 @@ import {
   prefixLength,
   InvalidKeyHashError,
   InvalidPublicKeyError,
+  ValidationResult,
   invalidErrorDetail,
 } from '@taquito/utils';
 import {
@@ -29,7 +30,7 @@ import {
 } from './michelson/codec';
 import { Uint8ArrayConsumer } from './uint8array-consumer';
 import { pad } from './utils';
-import { InvalidAddressError, InvalidContractAddressError, ValidationResult } from '@taquito/core';
+import { InvalidAddressError, InvalidContractAddressError } from '@taquito/core';
 
 // https://tezos.gitlab.io/shell/p2p_api.html specifies data types and structure for forging
