+++ packages/taquito-local-forging/src/taquito-local-forging.ts
@@ -8,11 +8,11 @@ import { CODEC } from './constants';
 import { decoders } from './decoder';
 import { encoders } from './encoder';
 import { Uint8ArrayConsumer } from './uint8array-consumer';
-import { invalidErrorDetail, validateBlock } from '@taquito/utils';
+import { validateBlock, ValidationResult, invalidErrorDetail } from '@taquito/utils';
 import { InvalidOperationSchemaError } from './error';
 import { validateMissingProperty, validateOperationKind } from './validator';
 import { ProtocolsHash } from './protocols';
-import { InvalidBlockHashError, InvalidOperationKindError, ValidationResult } from '@taquito/core';
+import { InvalidBlockHashError, InvalidOperationKindError } from '@taquito/core';
 
 export { CODEC, opMapping, opMappingReverse } from './constants';
 export * from './decoder';
