+++ packages/taquito-michelson-encoder/src/tokens/comparable/key_hash.ts
@@ -5,9 +5,8 @@ import {
   TokenValidationError,
   SemanticEncoding,
 } from '../token';
-import { encodeKeyHash, validateKeyHash } from '@taquito/utils';
+import { encodeKeyHash, validateKeyHash, ValidationResult } from '@taquito/utils';
 import { BaseTokenSchema } from '../../schema/types';
-import { ValidationResult } from '@taquito/core';
 
 export class KeyHashValidationError extends TokenValidationError {
   name = 'KeyHashValidationError';
