+++ packages/taquito-michelson-encoder/src/tokens/comparable/tx_rollup_l2_address.ts
@@ -1,5 +1,10 @@
 import { BaseTokenSchema } from '../../schema/types';
-import { b58decodeL2Address, encodeL2Address, validateAddress } from '@taquito/utils';
+import {
+  b58decodeL2Address,
+  encodeL2Address,
+  validateAddress,
+  ValidationResult,
+} from '@taquito/utils';
 import {
   ComparableToken,
   SemanticEncoding,
@@ -7,7 +12,6 @@ import {
   TokenFactory,
   TokenValidationError,
 } from '../token';
-import { ValidationResult } from '@taquito/core';
 
 export class TxRollupL2AddressValidationError extends TokenValidationError {
   name = 'TxRollupL2AddressValidationError';
