+++ packages/taquito-utils/src/verify-signature.ts
@@ -9,15 +9,11 @@ import {
   prefix,
   validatePublicKey,
   validateSignature,
+  ValidationResult,
 } from './taquito-utils';
 import elliptic from 'elliptic';
 import toBuffer from 'typedarray-to-buffer';
-import {
-  InvalidPublicKeyError,
-  InvalidMessageError,
-  InvalidSignatureError,
-  ValidationResult,
-} from '@taquito/core';
+import { InvalidPublicKeyError, InvalidMessageError, InvalidSignatureError } from '@taquito/core';
 
 type PkPrefix = Prefix.EDPK | Prefix.SPPK | Prefix.P2PK | Prefix.BLPK;
 type SigPrefix = Prefix.EDSIG | Prefix.SPSIG | Prefix.P2SIG | Prefix.SIG;
