+++ packages/taquito-utils/test/validators.spec.ts
@@ -1,5 +1,6 @@
 import {
   validateAddress,
+  ValidationResult,
   validateChain,
   validateKeyHash,
   validateContractAddress,
@@ -9,7 +10,6 @@ import {
   validateProtocol,
   validateBlock,
 } from '../src/validators';
-import { ValidationResult } from '@taquito/core';
 
 describe('validateAddress', () => {
   it('Validate address properly', () => {
